refactor: organizes FormTags exports in their respective folder
