+++ src/components/ui/form-tags/index.ts
@@ -1,0 +1,2 @@
+export { default as FormTags } from './FormTags.vue';
+export type { FormTagsProps, SettingsInterface } from './types';
